+++ packages/core/src/Provider/Provider.ts
@@ -1,7 +1,8 @@
 import { Cardano } from '..';
-import { Percent } from '../Cardano';
+
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Logger } from 'ts-log';
+import { Percent } from '@cardano-sdk/util';
 import { Tip } from '@cardano-ogmios/schema';
 
 export type HealthCheckResponse = {