+++ packages/util-dev/src/mockProviders/mockChainHistoryProvider.ts
@@ -1,6 +1,6 @@
 import * as AssetId from '../assetId';
 import { Cardano, Paginated } from '@cardano-sdk/core';
-import { currentEpoch, ledgerTip, stakeKeyHash } from './mockData';
+import { currentEpoch, handleAssetId, ledgerTip, stakeKeyHash } from './mockData';
 import { somePartialStakePools } from '../createStubStakePoolProvider';
 import delay from 'delay';
 
@@ -129,7 +129,13 @@ export const queryTransactionsResult: Paginated<Cardano.HydratedTx> = {
             address: Cardano.PaymentAddress(
               'addr_test1qq585l3hyxgj3nas2v3xymd23vvartfhceme6gv98aaeg9muzcjqw982pcftgx53fu5527z2cj2tkx2h8ux2vxsg475q2g7k3g'
             ),
-            value: { assets: new Map([[AssetId.TSLA, 1n]]), coins: 5_000_000n }
+            value: {
+              assets: new Map([
+                [AssetId.TSLA, 1n],
+                [handleAssetId, 1n]
+              ]),
+              coins: 5_000_000n
+            }
           }
         ],
         validityInterval: {