+++ packages/wallet/test/services/BalanceTracker.test.ts
@@ -3,7 +3,7 @@
 /* eslint-disable space-in-parens */
 import { BehaviorObservable } from '@cardano-sdk/util-rxjs';
 import { Cardano, coalesceValueQuantities } from '@cardano-sdk/core';
-import { DelegationTracker, createBalanceTracker } from '../../src/services';
+import { DelegationTracker, createBalanceTracker, createUtxoBalanceByAddressTracker } from '../../src/services';
 import { createTestScheduler, mockProviders } from '@cardano-sdk/util-dev';
 
 const { utxo, utxo2 } = mockProviders;
@@ -44,6 +44,67 @@ describe('createBalanceTracker', () => {
       });
       expectObservable(balanceTracker.utxo.unspendable$).toBe('-a-----', {
         a: coalesceValueQuantities(utxo2.map((u) => u[1].value))
+      });
+    });
+  });
+
+  it('createUtxoBalanceByAddressTracker returns balance filtered by addresses', () => {
+    const address = utxo[0][1].address;
+    const address2 = Cardano.PaymentAddress('addr_test1qzs0umu0s2ammmpw0hea0w2crtcymdjvvlqngpgqy76gpfnuzcjqw982pcftgx53fu5527z2cj2tkx2h8ux2vxsg475qp3y3vz');
+    const utxoMultipleAddresses = utxo.slice(0, 3);
+
+    // Change the first utxo address so it's filtered out
+    utxoMultipleAddresses[0][1].address = address2;
+    createTestScheduler().run(({ hot, expectObservable }) => {
+      const available$ = hot(  'aa', { a: utxoMultipleAddresses }) as unknown as BehaviorObservable<Cardano.Utxo[]>;
+      const total$ = hot(      'aa', { a: utxoMultipleAddresses }) as unknown as BehaviorObservable<Cardano.Utxo[]>;
+      const unspendable$ = hot('aa', { a: utxoMultipleAddresses }) as unknown as BehaviorObservable<Cardano.Utxo[]>;
+
+      const balanceByAddressTracker = createUtxoBalanceByAddressTracker(
+        { available$, total$, unspendable$ },
+        [address]
+      );
+
+      const balanceBy2AddressesTracker = createUtxoBalanceByAddressTracker(
+        { available$, total$, unspendable$ },
+        [address, address2]
+      );
+
+      const balanceNoFilterTracker = createUtxoBalanceByAddressTracker({ available$, total$, unspendable$ });
+
+
+      // Check emitted values when filtering by one address
+      const filteredUtxos = utxoMultipleAddresses.slice(1);
+      expectObservable(balanceByAddressTracker!.utxo.total$).toBe('a', {
+        a: coalesceValueQuantities(filteredUtxos.map((u) => u[1].value))
+      });
+      expectObservable(balanceByAddressTracker!.utxo.available$).toBe('a', {
+        a: coalesceValueQuantities(filteredUtxos.map((u) => u[1].value))
+      });
+      expectObservable(balanceByAddressTracker!.utxo.unspendable$).toBe('a', {
+        a: coalesceValueQuantities(filteredUtxos.map((u) => u[1].value))
+      });
+
+      // Check emitted values when there is no filtering, so all utxos are returned
+      expectObservable(balanceNoFilterTracker!.utxo.total$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
+      });
+      expectObservable(balanceNoFilterTracker!.utxo.available$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
+      });
+      expectObservable(balanceNoFilterTracker!.utxo.unspendable$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
+      });
+
+      // Check emitted values when there filtering by two addresses
+      expectObservable(balanceBy2AddressesTracker!.utxo.total$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
+      });
+      expectObservable(balanceBy2AddressesTracker!.utxo.available$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
+      });
+      expectObservable(balanceBy2AddressesTracker!.utxo.unspendable$).toBe('a', {
+        a: coalesceValueQuantities(utxoMultipleAddresses.map((u) => u[1].value))
       });
     });
   });